finished BreakPoint UI
Tried to get dynamic layout to work, but egui is not yet mature
enough to facilitate this without great effort, so the layout of version
0.1.0 is to be completely static. If a mobile version is added, 2 layout
varients can exist. I may need to write a pull request for egui at some
point, but i don't have time right now.

BreakPointBuilder is no more. I might well add it again later, but it
does not belong in the MVP. only BreakPoint exists.