added Vec2 from BreakPoint.
next task is to add a trait extention for Painter that allow breakpoint
to draw on an already-allocated rect. This will follow the same pattern
as `Widget`'s `fn ui(&mut self, ui: &mut Ui) -> Response`, but it will
take a rect instead of a Ui, and not return anything.