regex is now in the project for parseing
dragvalues should from now on have custom parsing and formatting fns.
this is easy to do with regex. has been done on slope.

TODO:
the length of a line is calculated incorrectly, so that it aligns to the
next point's horizontal positon, rather than vertical position as it
should.

change slope so that 0.0 represents flat ground and infinity is a wall,
this is mathmatically backwards, but i think probably more intuative.
mak e a new branch for this tho cuz it might not be better.