plotter development.
limit is now plotted (it wasn't before)

slope angles are project to the offset of next point (execpt in the case
of limit, where there is not next point). The elevation of the next
point is stored independently, so it will be visably obvioyus if they do
not allign. furthrmore, the next commit will feature a validator method
on slopestake that determins if it's points are valid in relation to
eachother. this will be used to make offending points red or something.