Everything is (de)serializable!
everything is also a mess.

Eq and hash are implemneted manually on Station, allowing it to be used
as a key in a hashmap, however, the program will panic if a NaN station
ever exists. This (and elevation and offset) should really be fixed
point ints. that is coming soon.